Fixed issue with querying vectors
resolved issue across docs, links, and videos. Shortened some functions, removed miscellaneous functions, made some more efficient. Still need to resolve inconsistency with similarity search with certain small content or vague query.

+++ routes.py
@@ -1,4 +1,4 @@
-from flask import render_template, request, Flask, make_response
+from flask import render_template, request
 from flask_wtf import FlaskForm
 from wtforms import StringField, SubmitField
 from werkzeug.utils import secure_filename
@@ -11,11 +11,9 @@ from pytube import YouTube
 from flask_limiter import Limiter
 import whisper
 import pinecone
-import docx2txt
-import mimetypes
 from flask_wtf.csrf import CSRFProtect
 from app import app, UPLOAD_FOLDER, PINECONE_API_KEY, PINECONE_API_ENV
-from utilities import allowed_file, read_content, index_pdf, perform_query, extract_text_from_pdf, limit_tokens
+from utilities import allowed_file, read_content, index_pdf, perform_query, limit_tokens
 
 app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY', config.SECRET_KEY)
 csrf = CSRFProtect(app)
@@ -58,9 +56,12 @@ def search():
                 file_extension = os.path.splitext(secure_filename(pdf.filename))[1].lower()
                 temp_path = os.path.join(tempfile.gettempdir(), secure_filename(pdf.filename))
                 pdf.save(temp_path)
+                app.logger.debug("TEMPFILE PDF CONTENT:" + pdf.content_type)
+                app.logger.debug("TEMPFILE PDF PATH:" + temp_path)
 
                 if file_extension in ['.pdf', '.txt', '.docx']:
                     pdf_content = read_content(temp_path)
+                    app.logger.debug("PDF CONTENT:" + pdf_content)
                 else:
                     return "Unsupported file format."
 
@@ -77,7 +78,7 @@ def search():
             app.logger.error(f"Error encountered during search: {str(e)}")
             app.logger.error(traceback.format_exc())
             return render_template('error.html', error_message=str(e))
-
+        
 @app.route('/index_from_link', methods=['POST'])
 def index_from_link():
     form = MyForm()
@@ -89,48 +90,43 @@ def index_from_link():
             app.logger.debug(f"Received link: {link} and query: {query}")
 
             if link and query:
-                temp_file = tempfile.NamedTemporaryFile(delete=False)
+                response = requests.get(link)
+                app.logger.debug(f"HTTP response status: {response.status_code}")
+                if response.status_code != 200:
+                    return "Failed to fetch content from the provided link."
+                
+                # Write response content to a temporary file
+                temp_file = tempfile.NamedTemporaryFile(delete=False, suffix=os.path.splitext(link)[1])
                 temp_file_path = temp_file.name
-                try:
-                    response = requests.get(link)
-                    app.logger.debug(f"HTTP response status: {response.status_code}")
-                    if response.status_code == 200:
-                        temp_file.write(response.content)
-                    else:
-                        return "Failed to fetch content from the provided link."
-                    mime_type, _ = mimetypes.guess_type(link)
-                    app.logger.debug(f"Determined MIME type: {mime_type}")
-                    if mime_type == 'application/pdf':
-                        content_text = extract_text_from_pdf(temp_file_path)
-                        content_text = limit_tokens(content_text)
-                        app.logger.debug(f"Extracted content type: {type(content_text)}")
-                    elif mime_type == 'text/plain':
-                        with open(temp_file_path, 'r', encoding='utf-8') as txt_file:
-                            content_text = txt_file.read()
-                            content_text = limit_tokens(content_text)
-                            app.logger.debug(f"Extracted content type: {type(content_text)}")
-                    elif mime_type == 'application/vnd.openxmlformats-officedocument.wordprocessingml.document':
-                        docx_content = temp_file.read()
-                        content_text = docx2txt.process(BytesIO(docx_content))
-                        content_text = limit_tokens(content_text)
-                        app.logger.debug(f"Extracted content type: {type(content_text)}")
-                    else:
-                        return "Unsupported link format."
-                    all_results = []
-                    for text_chunk in content_text:
-                        index = index_pdf(text_chunk)
-                        app.logger.debug(f"Index object created: {index}")
-                        results = perform_query(index, query)
-                        all_results.extend(results)
-                    return render_template('search_results.html', query=query, results=results)
-                finally:
-                    temp_file.close()
-                    os.unlink(temp_file_path)
+                temp_file.write(response.content)
+                temp_file.close()
+                app.logger.debug("TEMPFILE CONTENT:" + link)
+                app.logger.debug("TEMPFILE PATH:" + temp_file_path)
+
+                file_extension = os.path.splitext(link)[1].lower()
+                if file_extension in ['.pdf', '.txt', '.docx']:
+                    pdf_content = read_content(temp_file_path)
+                    app.logger.debug("PDF CONTENT:" + pdf_content)
+                else:
+                    os.unlink(temp_file_path)  # Clean up temp file
+                    return "Unsupported link format."
+
+                if pdf_content is None:
+                    os.unlink(temp_file_path)  # Clean up temp file
+                    return "Error processing the file."
+
+                index = index_pdf(pdf_content)
+                results = perform_query(index, query)
+                os.unlink(temp_file_path)  # Clean up temp file after processing
+                return render_template('search_results.html', query=query, results=results)
+            
             return "Please provide both a link and a query."
         except Exception as e:
+            import traceback
             app.logger.error(f"index_from_link: Error encountered - {str(e)}")
+            app.logger.error(traceback.format_exc())
             return render_template('error.html', error_message=str(e))
-
+        
 @app.route('/video_transcribe', methods=['POST'])
 def video_transcribe():
     form = MyForm()
@@ -146,24 +142,41 @@ def video_transcribe():
                     audio_stream = yt.streams.filter(only_audio=True).first()
                     audio_stream.download(output_path=temp_dir.name, filename='youtube_audio_copy.mp3')
                     app.logger.debug("Audio downloaded successfully.")
+                except Exception as download_error:
+                    app.logger.error(f"Error downloading: {download_error}")
+                    return render_template('error.html', error_message=str(download_error))
+
+                try:
                     result = model.transcribe(temp_audio_mp3_path, fp16=False)
                     transcribed_text = result["text"]
                     transcribed_texts = limit_tokens(transcribed_text)
                     for text in transcribed_texts:
-                        index = index_pdf(text)
-                        results = perform_query(index, query)
+                        aggregate_text = " ".join(text)
                     app.logger.debug(transcribed_text)
-                    index = index_pdf(transcribed_text)
+                except Exception as read_error:
+                    app.logger.error(f"Error reading: {read_error}")
+                    return render_template('error.html', error_message=str(read_error))
+
+                try:
+                    index = index_pdf(aggregate_text)
                     app.logger.debug("Indexing successful")
+                except Exception as index_error:
+                    app.logger.error(f"Error indexing: {index_error}")
+                    return render_template('error.html', error_message=str(index_error))
+
+                try:
                     results = perform_query(index, query)
                     app.logger.debug("Query successful")
                     return render_template('search_results.html', query=query, results=results)
-                except Exception as download_error:
-                    app.logger.error(f"Error downloading: {download_error}")
-                    return f"Error downloading the YouTube video: {str(download_error)}"
+
+                except Exception as query_error:
+                    app.logger.error(f"Error querying: {query_error}")
+                    return render_template('error.html', error_message=str(query_error))
+
                 finally:
                     if os.path.exists(temp_audio_mp3_path):
                         os.remove(temp_audio_mp3_path)
+
         except Exception as e:
             app.logger.error(f"Error in video_transcribe: {e}")
             return render_template('error.html', error_message=str(e))
